Rename ReviewVideoRecorder*
