+++ app/javascript/components/ReviewForm/ReviewVideoRecorderCommon.tsx
